requirements.txt
llama-cpp-python
fastapi
uvicorn
sse-starlette
pydantic